+++ src/renderer_raster_clusters_lod.cpp
@@ -136,6 +136,8 @@ bool RendererRasterClustersLod::init(Resources& res, RenderScene& rscene, const 
     m_sceneBuildShaderio.renderClusterInfos =
         mem.append(sizeof(shaderio::ClusterInfo) * m_sceneBuildShaderio.maxRenderClusters, 8);
 
+    m_sceneBuildShaderio.instanceStates = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
+
     if(config.useSorting)
     {
       m_sceneBuildShaderio.instanceSortKeys   = mem.append(sizeof(uint32_t) * m_renderInstances.size(), 4);
@@ -146,6 +148,7 @@ bool RendererRasterClustersLod::init(Resources& res, RenderScene& rscene, const 
     m_resourceReservedUsage.operationsMemBytes += m_sceneDataBuffer.bufferSize;
 
     m_sceneBuildShaderio.renderClusterInfos += m_sceneDataBuffer.address;
+    m_sceneBuildShaderio.instanceStates += m_sceneDataBuffer.address;
     m_sceneBuildShaderio.instanceSortKeys += m_sceneDataBuffer.address;
     m_sceneBuildShaderio.instanceSortValues += m_sceneDataBuffer.address;